requirements.txt
sys
os
numpy
scipy.constants
tqdm
